Decided to move some of the files around so that the development files and production files are completely separate Also renamed the Grunt-generated files so it's easier to what they do.
This is the last change I am going to make. The next step I would have taken would be to implement the observer pattern so that the model can notify the view when its data changes, instead of having manual updates scattered through the controller.
However at present the code is in a fairly good and working condition and I decided to stop before getting stuck into a large addition.